+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.5.0] - YYYY-MM-DD
 
+- [#3 - Improve interceptor interfaces](https://gitlab.com/yggdrasilts/axiosfit/issues/3)
+
 ## [0.4.4] - 2019-10-27
 
 ### Bugfixes